Restructured httpConfigs to avoid circular referencing
