+++ src/main/java/uk/gov/cabinetoffice/csl/client/identity/IdentityClient.java
@@ -1,14 +1,16 @@
 package uk.gov.cabinetoffice.csl.client.identity;
 
+import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Qualifier;
 import org.springframework.http.RequestEntity;
-import org.springframework.stereotype.Service;
+import org.springframework.stereotype.Component;
 import uk.gov.cabinetoffice.csl.client.IHttpClient;
 import uk.gov.cabinetoffice.csl.domain.identity.OAuthToken;
 
 import java.util.Collections;
 
-@Service
+@Component
+@Slf4j
 public class IdentityClient implements IIdentityClient {
 
     private final IHttpClient client;
@@ -19,6 +21,7 @@ public class IdentityClient implements IIdentityClient {
 
     @Override
     public OAuthToken getServiceToken() {
+        log.debug("Getting service token from identity service");
         RequestEntity<Void> request = RequestEntity
                 .post("/oauth/token", Collections.singletonMap("grant_type", "client_credentials"))
                 .build();
