+++ src/main/java/uk/gov/cabinetoffice/csl/client/learnerRecord/LearnerRecordClientConfig.java
@@ -1,0 +1,32 @@
+package uk.gov.cabinetoffice.csl.client.learnerRecord;
+
+import org.springframework.beans.factory.annotation.Value;
+import org.springframework.boot.web.client.RestTemplateBuilder;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+import org.springframework.web.client.RestTemplate;
+import uk.gov.cabinetoffice.csl.client.HttpClient;
+import uk.gov.cabinetoffice.csl.client.IHttpClient;
+import uk.gov.cabinetoffice.csl.service.auth.RestTemplateOAuthInterceptor;
+
+@Configuration
+public class LearnerRecordClientConfig {
+
+    @Value("${learnerRecord.serviceUrl}")
+    private String learnerRecordBaseUrl;
+
+    private final RestTemplateOAuthInterceptor restTemplateOAuthInterceptor;
+
+    public LearnerRecordClientConfig(RestTemplateOAuthInterceptor restTemplateOAuthInterceptor) {
+        this.restTemplateOAuthInterceptor = restTemplateOAuthInterceptor;
+    }
+
+    @Bean(name = "learnerRecordHttpClient")
+    IHttpClient learnerRecordClient(RestTemplateBuilder restTemplateBuilder) {
+        RestTemplate restTemplate = restTemplateBuilder
+                .rootUri(learnerRecordBaseUrl)
+                .additionalInterceptors(restTemplateOAuthInterceptor)
+                .build();
+        return new HttpClient(restTemplate);
+    }
+}
